call-ml-service/requirements.txt
pika==1.3.1
python-dotenv==1.0.1
assemblyai==0.48.4
google-genai==1.56.0
pydantic==2.7.1
